chore: Move examples tests to test folder
